refactor(audit): move the event types to the package that owns the vocabulary
AuditEventCommon, AuditResourceAction, AuditEvent and StoredAuditEvent were the
whole of packages/audit/lib/event.ts, and every import in it came from
@nangohq/types — the file was 56 lines of derivation living one package away
from what it derives from. AuditEventCommon was not even exported.

The reason it matters beyond tidiness: @nangohq/types cannot depend on
@nangohq/audit, so while AuditEvent lived there, api.ts was structurally unable
to reference it. Deriving ApiAuditTrailEvent from the same union is now
possible.

Consumers are unaffected: the audit barrel re-exports all four, so importing
them from @nangohq/audit still works.
